refactor: rename mongodb file to utils
