+++ app/page.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react"
 import Cookies from "js-cookie"
 
 import { Button } from "./components/Button"
-import { handleClick } from "./mongodb/temporaryHandleClick"
+import { handleClick } from "./utils/temporaryHandleClick"
 import { TranscriptCard } from "./components/TranscriptCard"
 import { SelectedTranscript } from "./components/SelectedTranscript"
 
@@ -18,10 +18,11 @@ export default function Home() {
   return (
     <main className="flex min-h-screen flex-row items-center justify-between p-16">
       <div className="flex flex-col">
-        <Button text={""} onClick={handleClick} />
+        <Button text={"Get notes"} onClick={handleClick} />
         <TranscriptCard />
       </div>
       <SelectedTranscript />
+      <Button text={"Log out"} onClick={handleClick} />
     </main>
   )
 }
